fix: SOS-25 so:inputattribute was not working. the safe-online-taglib jar must be in the same place as the facelets jar (classloading stuff)
WARNING: base-distro has been updated so replace your local jboss with it.
WARNING: one of our artifacts is now in JBoss so outside the ear. Remember this when you are working on safe-online-taglib and your changes are not coming through. you need to update your jboss install.

fix: SOS-26 easier captcha engine



git-svn-id: svn+ssh://buildserver.lin-k.net/scm/svn/products/safe-online/trunk@1118 cd7dd898-ab41-4057-98bc-1513970a52e4