Cinema Demo now works completely.
rem: remove useless attach() attempts in CinemaSession.  The resulting attached entity is detached again anyway because CinemaSession is not in EJB context.
add: logout button and redo style of user info table a little.
fix: ticket history time stamp formatting as date, not as currency.
fix: duration is in seconds but we were adding it to a timestamp in milliseconds.  Added *1000.
add: log noise.
fix: theatre and film strings swapped.
fix: attach the user entity given to us by wicket.  this user entity is detached; before being able to modify it we need to get the attached entity.
fix: javadoc warning in SafeOnlineLoginUtils

git-svn-id: svn+ssh://buildserver.lin-k.net/scm/svn/products/safe-online/trunk@1482 cd7dd898-ab41-4057-98bc-1513970a52e4