Fix startable dependencies.
fix: startables had messy dependencies; Startable class was in s-o-model; which does not belong there.  Moved s-o-startup (which is a war that just runs all the startables) to s-o-startup-runtime and created s-o-startup as a lib that now contains the Startable from s-o-model.  Now beid,digipas,demo,... depend on s-o-startable and s-o-startable-runtime and no longer on s-o-model.