Deployment fixes.
fix: context root path of deployment descriptors. Maven puts a contextroot element in application.xml with the name of the artifact; which is crap.  TODO: Find a way to just tell maven NOT to put a contextroot element in application.xml so we can get rid of the modules section in the *-deploy artifacts' pom.xml files.
fix: deploy digipass before demo by putting a silly 1 in its name.
add: zip generating assembly descriptors.