Fix deployment again.
fix: regression when s-o-startup-runtime was added back to s-o-model.  It shouldn't go there because then maven freaks out and adds s-o-startup-runtime to artifacts that depend on s-o-model-client even when s-o-startup-runtime is excluded explicitly.
fix: web context paths start with a /.
upd: use maven properties for web context paths; don't hardcode (they are used elsewhere in the code too and need to be kept in sync).