Use correct application names for cinema & payment demos. Rename ear files for better deployment ordering.
upd: It's safer and more consistent to force demo to deploy as last than it is to force certain ears (those that start with a letter that sorts after D) to deploy before demo.  Get rid of the 1's in the ear filenames and just put a Z in the demo ear filename.