Unify XML URI resolution in a shared catalog and fix remote accessing of schemas.
fix: wsimport/xjc are bugged.  When resolving URIs from schemas linked by the schema/wsdl it's parsing, it doesn't seem to use the catalog anymore, which breaks local-only resolution of the schema files.  Workaround is to make all refs to schemalocations local.  WsImport *still* complains, even though it finds and successfully parses all the catalogs it needs (shows a WARNING about not being able to find a schema in the resources dir that it managed to load successfully earlier from the catalog dir).