Template adjustments to fix odd Help link.
upd: sidebar template now takes an argument, which is the message that
goes into the sidebar.  Since this message is an essential part of the
sidebar, it really belonged in the sidebar border itself.
upd: since the message has been extracted; the only part that needs
overriding for the sidebar are now optional additional links.  that
means we can now use a single <ul> and let the child add <li>s to it.
fix: this fixes an issue where the help link which was defined in the
sidebar border was spaced away from the additional links which came
below it in a different <ul>.  They are now in the same <ul>.