fix SOS-396 authority key identifier
missing authority key identifier were accepted, but there was a bug in finding
the trust point.

Also, we now accept non CA certificates in the trust path as long as the
certificate that needs validation is the non CA certificate. In other words:
you now can add leave certificates in the trust domain