Lots of context param refactoring.
upd: lots of context parameters that used to be URLs are now paths inside their respective webapps and use SafeOnlineConfig to resolve them into absolute URLs.
ren: renamed the two LandingServlet's in device-sdk to AbstractDeviceAuthenticationLandingServlet and AbstractDeviceManagementLandingServlet.
ren: renamed all that implemented those servlets to DeviceAuthenticationLandingServlet and DeviceManagementLandingServlet
mov: moved the latter from device-sdk into its own package just like auth is so that it's on the same level.
upd: lots of javadoc updates so that all those web.xml parameters are properly documented.

** We should always use {@link #SOME_CONTEXT_PARAM} in the javadoc of
methods that require that context param to be set.  Then, add proper
Javadoc to that constant field which explains in all detail and with
clear and unambiguous language what the parameter does and what its
values are expected to be.
That way, it's always clear when refering to code later on what it uses
and how those parameters are to be interpreted: especially important for
the SDK which will also be used by outside developers! ***