More URL->Path refactoring.
upd: more Url -> Path and absolutification of those paths based on the hostname and webapp url configuration.
upd: lots of javadoc updates so that we know what we're dealing with.
add: SafeOnlineConfig and WebappPath to web.xmls.  These are used to determine what the absolute URL for the paths is going to be.