A.. lot of changes, just read the full description.
rem: safe-online.version removed from all pom files and safeonline version explicitly mentioned for all safeonline artifacts: fixes some oddness when artifacts with a different version depend on artifacts that use safe-online.version for its version.  (Eg, siemens-*bundle (1.0-SNAPSHOT depending transitively on safe-online-container-jsf will make maven think we need version 1.0-SNAPSHOT of the latter instead of whatever safe-online.version says.)
rem: -SNAPSHOT from Import-Package declarations of bundles.
add: safe-online-model-node:  This artifact is a simple EJB artifact that contains the SafeOnlineNodeStartable bean which loads the SafeOnlineNodeKeystore configuration into the database.  Put it in a -model-* artifact to conform with all the other Startable beans we have.
add: NodeUtils utility class which allows us to make easy requests on the local node; such as which one it is, what its absolute endpoint URL is, etc.  Very handy for all web applications that run in OLAS nodes (such as our external devices) which need to know the URL of their node.
upd: AbstractInjectionServlet now wraps HttpServletRequest and ~Response by default when 'appbase' is defined in the SafeOnlineConfig.  This is used to avoid validation and redirection problems when the server is behind a reverse proxy or load balancer.  The wrapper changes the requestURI and the way relative redirects are resolved to absolute ones using the URL base of the load balancer/proxy instead of that of the JBoss server.  This base URL is configured in 'appbase' in SafeOnlineConfig.
add: AbstractNodeInjectonServlet and AbstractLandingInjectionServlet.  These are variants of AbstractInjectionServlet which override the getWrappedEndpoint method and respectively return the endpoint configured in the local node and the endpoint set in SafeOnlineConfig for application landing servlets.
ren: Renamed /device/auth etc to /device-auth etc.  Currently the way the proxy-url-fixing works by relying on a single-element path inside the webapp; and there really isn't any reason why we would want to complicate that for this situation; at least not for now.
ren: Renamed a lot of servlets to better describe what they are.  Such as DeviceLandingServlet -> DeviceAuthenticationLandingServlet (to be consistent with DeviceRegistrationLandingServlet).
add: Several servlets in Device-SDK are now abstract and need to be implemented by the device provider.  This is because the device provider needs to specify what URL to use for wrapping the servlet request.  Normally, the device provider will want to use the local node's location for this purpose; but since we need to keep Device-SDK Node-Independant we need to make device-providers implement this logic themselves by overriding getWrappedEndpoint().  They just return the local node's location there.
fix: lots of fixes on which base URL to use for converting a path as obtained from web.xml into a fully absolute URL.  The way we want to do this depends on whether the path refers to within some (demo) application, or refers to within an OLAS node, or refers to a landing servlet.
fix: correct some javadoc.
fix: correct some logic in the HttpServletEndpoint*Wrapper for rebasing urls or converting relative urls into absolute ones.