Last hand at externalizing host parameters.
fix: JAAS filter needs to trigger before wicket.
rem: we really don't need HttpServletRequest to get the SafeOnlineConfig.  Take it out of web.xml and hardcode the location (which is always in jboss' conf dir) in the SafeOnlineConfig class.  Then we can access the config outside servlet-context (which we need to do for the startables).  This simplifies a lot of other uses of the config, too - clean that up.
add: node configuration to safeonlineconfig.
upd: startables now use safeonlineconfig to obtain their node's initial configuration.
upd: made convenience methods of safeonlineconfig static for more convenience.
upd: no need to cache safeonlineconfig in a map: there is only one.